test(graphics): save snapshots for RLE test
Instead of directly depending on the unmaintained rdp-rs to decompress
RLE tiles dynamically, a few snapshots are saved to test against.

The benefit is to remove a lot of outdated dependencies from our
(developement) dependency graph.

+++ crates/ironrdp-session/src/fast_path.rs
@@ -128,9 +128,9 @@ impl Processor {
                             match ironrdp_graphics::rle::decompress(
                                 update.bitmap_data,
                                 &mut buf,
-                                update.width,
-                                update.height,
-                                update.bits_per_pixel,
+                                usize::from(update.width),
+                                usize::from(update.height),
+                                usize::from(update.bits_per_pixel),
                             ) {
                                 Ok(RlePixelFormat::Rgb16) => image.apply_rgb16_bitmap(&buf, &update.rectangle)?,
